requirements.txt
torch>=2.1.0
transformers>=4.44.0
peft>=0.13.0
datasets>=3.0.0
accelerate>=1.0.0
Pillow>=10.0.0
matplotlib>=3.8.0
tqdm>=4.66.0
